+++ src/main/java/consultorio/api/controller/HistoriaClinicaController.java
@@ -85,7 +85,7 @@ public class HistoriaClinicaController {
                     return gson.toJson(Map.of("error", "Paciente y profesional son obligatorios"));
                 }
 
-                Paciente p = pacienteDAO.buscarPorId(Math.toIntExact(input.pacienteId));
+                Paciente p = pacienteDAO.buscarPorId((long) Math.toIntExact(input.pacienteId));
                 ProfesionalSalud prof = profDAO.buscarPorId(Math.toIntExact(input.profesionalId));
 
                 if (p == null || prof == null) {
